Move validate tests out of integration
