+++ jetstream/tests/integration/test_config_integration.py
@@ -7,13 +7,11 @@ import pytest
 import pytz
 import toml
 from metric_config_parser.analysis import AnalysisSpec
-from metric_config_parser.config import Config, ConfigCollection, DefaultConfig, Outcome
+from metric_config_parser.config import Config, ConfigCollection
 from metric_config_parser.experiment import BucketConfig, Channel, Experiment
 from metric_config_parser.metric import AnalysisPeriod
-from metric_config_parser.outcome import OutcomeSpec
 
-from jetstream.config import ConfigLoader, validate
-from jetstream.dryrun import DryRunFailedError
+from jetstream.config import ConfigLoader
 from jetstream.statistics import LinearModelMean, Summary
 
 TEST_DIR = Path(__file__).parent.parent
@@ -134,104 +132,6 @@ class TestConfigIntegration:
             last_modified=datetime.datetime.now(),
         )
         extern.validate(ConfigLoader.configs, experiments[0])
-
-    def test_busted_config_fails(self, experiments):
-        config = dedent(
-            """\
-            [metrics]
-            weekly = ["bogus_metric"]
-
-            [metrics.bogus_metric]
-            select_expression = "SUM(fake_column)"
-            data_source = "clients_daily"
-            statistics = { bootstrap_mean = {} }
-            """
-        )
-        spec = AnalysisSpec.from_dict(toml.loads(config))
-        extern = Config(
-            slug="bad_experiment",
-            spec=spec,
-            last_modified=datetime.datetime.now(),
-        )
-        with pytest.raises(DryRunFailedError):
-            validate(extern, experiments[0])
-
-    def test_valid_outcome_validates(self):
-        config = dedent(
-            """\
-            friendly_name = "Fred"
-            description = "Just your average paleolithic dad."
-
-            [metrics.rocks_mined]
-            select_expression = "COALESCE(SUM(pings_aggregated_by_this_row), 0)"
-            data_source = "clients_daily"
-            statistics = { bootstrap_mean = {} }
-            friendly_name = "Rocks mined"
-            description = "Number of rocks mined at the quarry"
-            """
-        )
-        spec = OutcomeSpec.from_dict(toml.loads(config))
-        extern = Outcome(
-            slug="good_outcome",
-            spec=spec,
-            platform="firefox_desktop",
-            commit_hash="0000000",
-        )
-
-        validate(extern)
-
-    def test_busted_outcome_fails(self):
-        config = dedent(
-            """\
-            friendly_name = "Fred"
-            description = "Just your average paleolithic dad."
-
-            [metrics.rocks_mined]
-            select_expression = "COALESCE(SUM(fake_column_whoop_whoop), 0)"
-            data_source = "clients_daily"
-            statistics = { bootstrap_mean = {} }
-            friendly_name = "Rocks mined"
-            description = "Number of rocks mined at the quarry"
-            """
-        )
-        spec = OutcomeSpec.from_dict(toml.loads(config))
-        extern = Outcome(
-            slug="bogus_outcome",
-            spec=spec,
-            platform="firefox_desktop",
-            commit_hash="0000000",
-        )
-        with pytest.raises(DryRunFailedError):
-            validate(extern)
-
-    def test_valid_default_config_validates(self):
-        extern = DefaultConfig(
-            slug="firefox_desktop",
-            spec=self.spec,
-            last_modified=datetime.datetime.now(),
-        )
-        validate(extern)
-
-    def test_busted_default_config_fails(self):
-        config = dedent(
-            """\
-            [metrics]
-            weekly = ["bogus_metric"]
-
-            [metrics.bogus_metric]
-            select_expression = "SUM(fake_column)"
-            data_source = "clients_daily"
-            statistics = { bootstrap_mean = {} }
-            """
-        )
-        spec = AnalysisSpec.from_dict(toml.loads(config))
-        extern = DefaultConfig(
-            slug="firefox_desktop",
-            spec=spec,
-            last_modified=datetime.datetime.now(),
-        )
-        with pytest.raises(DryRunFailedError):
-            validate(extern)
 
     @pytest.mark.parametrize(
         "randomization_unit",
